Enable the setting for “key”
To get the key from the Fader component into the sequence generator,
I moved the Settings component into the Game component. When the
setting is changed, the handler in Game modifies the key value directly
on the sequence generator. I also disabled the other settings to make
clear what works and what doesn’t.

Now that we have different keys working, playing notes in the right
octave becomes a little more difficult. I added a sameNote function to
treat notes of the same pitch class as the same, even if they’re not in
the same octave.

Unfortunately, the initial chord that plays before the sequence starts
playing is still hard-coded to C major. In order to change this, I
should probably make the key part of the game’s state and turn
SequenceGenerator into a component that inherits it as a prop.