Allow manual replay of sequences
The flow here is based on the idea that, as long as the app is
recording the user’s notes, the sequence should be replayable. Up until
now, we were setting isRecording to true directly after calling playSequence, which meant we’d start recording before the sequence was
even started. Moving it to the finishSequence callback ensures that we
don’t start recording (and thus the sequence isn’t repayable) until the sequence is finished.

Synth.playSequence still has some hardcoded time values that we’ll need
to deal with later on.